User commit

+++ src/main/java/com/example/demo/repository/Studentrepo.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.Studententity.Studententity;
 
-public interface Studentrepo extends JpaRepository<Studententity, Long> {
+public interface Studentrepo extends JpaRepository<Studententity, Integer> {
 }
